requirements.txt
Django>=4.0.0
djangorestframework==3.13.1

flake8>=3.6.0,<3.7.0
python-decouple>=3.6